% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tf_idf.R
\name{bind_tf_idf}
\alias{bind_tf_idf}
\alias{bind_tf_idf_}
\title{Bind the term frequency and inverse document frequency of a tidy text
dataset to the dataset}
\usage{
bind_tf_idf(tbl, term_col, document_col, n_col)

bind_tf_idf_(tbl, term_col, document_col, n_col)
}
\arguments{
\item{tbl}{A tidy text dataset with one-row-per-term-per-document}

\item{term_col}{Column containing terms}

\item{document_col}{Column containing document IDs}

\item{n_col}{Column containing document-term counts}
}
\description{
Calculate and bind the term frequency and inverse document frequency of a
tidy text dataset, along with the product, tf-idf to the dataset. Each of
these values are added as columns.
}
\details{
\code{tf_idf} is given bare names, while \code{tf_idf_}
is given strings and is therefore suitable for programming with.

If the dataset is grouped, the groups are ignored but are
retained.

The dataset must have exactly one row per document-term combination
for this to work.
}
\examples{

library(dplyr)
library(janeaustenr)

book_words <- austen_books() \%>\%
  unnest_tokens(word, text) \%>\%
  count(book, word, sort = TRUE) \%>\%
  ungroup()

book_words

# find the words most distinctive to each document
book_words \%>\%
  bind_tf_idf(word, book, n) \%>\%
  arrange(desc(tf_idf))

}
