% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_casters.R
\name{cast_sparse}
\alias{cast_sparse}
\title{Create a sparse matrix from row names, column names, and values
in a table.}
\usage{
cast_sparse(data, row, column, value)
}
\arguments{
\item{data}{A tbl}

\item{row}{A bare column name to use as row names in sparse matrix}

\item{column}{A bare column name to use as column names in sparse matrix}

\item{value}{A bare column name to use as sparse matrix values, default 1}
}
\value{
A sparse Matrix object, with one row for each unique value in
the \code{row} column, one column for each unique value in the \code{column}
column, and with as many non-zero values as there are rows in \code{data}.
}
\description{
Create a sparse matrix from row names, column names, and values
in a table.
}
\examples{

dat <- data.frame(a = c("row1", "row1", "row2", "row2", "row2"),
                  b = c("col1", "col2", "col1", "col3", "col4"),
                  val = 1:5)

cast_sparse(dat, a, b)

cast_sparse(dat, a, b, val)

}

