% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_casters.R
\name{cast_sparse_}
\alias{cast_sparse_}
\title{Standard-evaluation version of cast_sparse}
\usage{
cast_sparse_(data, row_col, column_col, value_col = 1, ...)
}
\arguments{
\item{data}{A tbl}

\item{row_col}{String version of column to use as row names}

\item{column_col}{String version of column to use as column names}

\item{value_col}{String version of column to use as sparse matrix
values, or a numeric vector to use. Default 1 (to create a binary
matrix)}

\item{...}{Extra arguments to pass on to \code{\link{sparseMatrix}}}
}
\description{
Standard-evaluation version of cast_sparse
}
