% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_casters.R
\name{cast_tdm_}
\alias{cast_dtm}
\alias{cast_dtm_}
\alias{cast_tdm}
\alias{cast_tdm_}
\title{Casting a data frame to
a DocumentTermMatrix or a TermDocumentMatrix}
\usage{
cast_tdm_(data, term_col, document_col, value_col = 1,
  weighting = tm::weightTf, ...)

cast_tdm(data, term, document, value = 1, weighting = tm::weightTf, ...)

cast_dtm_(data, document_col, term_col, value_col = 1,
  weighting = tm::weightTf, ...)

cast_dtm(data, document, term, value = 1, weighting = tm::weightTf, ...)
}
\arguments{
\item{data}{Table with one-term-per-document-per-row}

\item{weighting}{The weighting function for the DTM/TDM
(default is term-frequency, effectively unweighted)}

\item{...}{Extra arguments passed on to
\code{\link{sparseMatrix}}}

\item{term, term_col}{(Bare) name of a column with terms}

\item{document, document_col}{(Bare) name of a column with documents}

\item{value, value_col}{(Bare) name of a column containing values}
}
\description{
This turns a "tidy" one-term-per-dopument-per-row data frame into a
DocumentTermMatrix or TermDocumentMatrix from the tm package. Each
can be called either with non-standard evaluation (bare column names)
or character vectors (for \code{cast_tdm_} and \code{cast_dtm_}).
}

