% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_tidiers.R
\name{dictionary_tidiers}
\alias{dictionary_tidiers}
\alias{tidy.dictionary}
\title{Tidy dictionary objects from the quanteda package}
\usage{
tidy.dictionary(x, regex = FALSE, ...)
}
\description{
Tidy dictionary objects from the quanteda package
}

