% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\name{get_quanteda_stopwords}
\alias{get_quanteda_stopwords}
\title{Get a tidy data frame of a single quanteda stop word lexicon}
\usage{
get_quanteda_stopwords(language = "english")
}
\arguments{
\item{language}{The language of the stop word lexicon specified, one of
\code{"english"}, \code{"SMART"}, \code{"danish"}, \code{"french"},
\code{"greek"}, \code{"hungarian"}, \code{"norwegian"}, \code{"russian"},
\code{"swedish"}, \code{"catalan"}, \code{"dutch"}, \code{"finnish"},
\code{"german"}, \code{"italian"}, \code{"portuguese"}, \code{"spanish"},
or \code{"arabic"}}
}
\value{
A tibble with two columns, \code{word} and \code{lexicon}. The
parameter \code{lexicon} is "quanteda" in this case.
}
\description{
Get a specific stop word lexicon from quanteda's \link[quanteda]{stopwords}
in a tidy format, with one row per word, in a form that can be joined or
anti-joined with a one-word-per-row dataset.
}
\examples{

library(dplyr)
get_quanteda_stopwords("spanish")
get_quanteda_stopwords("arabic")

}
