% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Get a tidy data frame of stopwords}
\usage{
get_stopwords(language = c("en", "fr"))
}
\arguments{
\item{language}{The stopwords language;
either "en" for english, or "fr" for french}
}
\value{
For english, a tbl_df with a \code{word} column, and a \code{lexicon}
column, which is either "onix", "SMART", or "snowball".

For french, a tbl_df with a single \code{word} column.
}
\description{
Get stopwords in a tidy format, with one row per word,
in a form that can be joined with a one-word-per-row dataset.
}
\examples{

get_stopwords("en")
get_stopwords("fr")

}
