% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glove_tidiers.R
\name{tidy.text2vec_glove_fit}
\alias{tidy.text2vec_glove_fit}
\title{Tidy glove word-to-vector fits}
\usage{
\method{tidy}{text2vec_glove_fit}(x, terms = NULL)
}
\arguments{
\item{x}{A text2vec_glove_fit object, such as returned from the
\code{\link[text2vec]{glove}} function.}

\item{...}{Extra arguments, not used}
}
\description{
Tidy glove word-to-vector fits from the text2vec package
}

