% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_words.R
\docType{data}
\name{nma_words}
\alias{nma_words}
\title{English negators, modals, and adverbs}
\format{
A data frame with 44 rows and 2 variables:
\describe{
\item{word}{An English word or bigram}
\item{modifier}{The modifier type for \code{word}, either "negator",
"modal", or "adverb"}
}
}
\source{
\url{http://saifmohammad.com/WebPages/SCL.html#NMA}
}
\usage{
nma_words
}
\description{
English negators, modals, and adverbs, as a data frame. A few of these
entries are two-word phrases instead of single words.
}
\keyword{datasets}
