% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_words.R
\name{stem_words}
\alias{stem_words}
\title{Word stemming}
\usage{
stem_words(df, col = word, language = "english")
}
\arguments{
\item{df}{the data.frame containing the unnested text}

\item{col}{the column containing the words. Default is word.}

\item{language}{language of the text. Default is english. The list of all supported languages
is available with \code{\link[SnowballC]{getStemLanguages}}}
}
\value{
a tibble with stemmed words
}
\description{
Word stemming
}
\examples{
library(janeaustenr)
library(dplyr)
original_books <- austen_books() \%>\%
 group_by(book) \%>\%
 mutate(linenumber = row_number()) \%>\%
 ungroup()

library(tidytext)
tidy_books <- original_books \%>\%
 unnest_tokens(word, text) \%>\%
 stem_words()

 tidy_books

}
