% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\docType{data}
\name{stop_words_foreign}
\alias{stop_words_foreign}
\title{Stop words from 56 languages}
\format{A list of 56 data frame with 1 variable each:
\describe{
 \item{word}{A stop word}
 }}
\source{
\itemize{
\item \url{https://github.com/stopwords-iso/stopwords-iso}
}
}
\usage{
stop_words_foreign
}
\description{
This list contains 56 data frames with stop words from various non-english
languages. You can acess them with \code{stop_words_foreign$} and the two letters
describing the language. For example \code{stop_words_foreign$fr} refer to french
stopwords. To find which initial refer to your language, please visit
\url{https://github.com/stopwords-iso}.
}
\keyword{datasets}
