% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{English stop words from TODO source}
\format{A data frame with 1149 rows and 2 variables:
\describe{
 \item{word}{An English word}
 \item{lexicon}{The source of the stop word. One of either TODO}
 }}
\source{
\url{TODO}
\url{TODO}
}
\usage{
stopwords
}
\description{
TODO
}
\keyword{datasets}

