% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{supported_dictionaries}
\alias{supported_dictionaries}
\title{Supported dictionaries}
\usage{
supported_dictionaries()
}
\description{
Get the hunspell dictionaries supported natively in \pkg{tidytext}.
These dictionaries were downloaded from \url{https://github.com/titoBouzout/Dictionaries}
}
