% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_tidiers.R
\name{tdm_tidiers}
\alias{tdm_tidiers}
\alias{tidy.DocumentTermMatrix}
\alias{tidy.TermDocumentMatrix}
\alias{tidy.dfm}
\alias{tidy.dfmSparse}
\alias{tidy.simple_triplet_matrix}
\title{Tidy DocumentTermMatrix, TermDocumentMatrix, and related objects
from the tm package}
\usage{
\method{tidy}{DocumentTermMatrix}(x, ...)

\method{tidy}{TermDocumentMatrix}(x, ...)

\method{tidy}{dfm}(x, ...)

\method{tidy}{dfmSparse}(x, ...)

\method{tidy}{simple_triplet_matrix}(x, row_names = NULL, col_names = NULL, ...)
}
\arguments{
\item{x}{A DocumentTermMatrix or TermDocumentMatrix object}

\item{...}{Extra arguments, not used}

\item{row_names}{Specify row names}

\item{col_names}{Specify column names}
}
\description{
Tidy a DocumentTermMatrix or TermDocumentMatrix into
a three-column data frame: \code{term{}}, and value (with
zeros missing), with one-row-per-term-per-document.
}
\examples{

if (requireNamespace("topicmodels", quietly = TRUE)) {
  data("AssociatedPress", package = "topicmodels")
  AssociatedPress

  tidy(AssociatedPress)
}

}
