% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_tidiers.R
\name{tidy_triplet}
\alias{tidy_triplet}
\title{Utility function to tidy a simple triplet matrix}
\usage{
tidy_triplet(x, triplets)
}
\arguments{
\item{x}{Object with rownames and colnames}

\item{triplets}{A data frame or list of i, j, x}
}
\description{
Utility function to tidy a simple triplet matrix
}

