% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytext.R
\docType{package}
\name{tidytext-package}
\alias{tidytext}
\alias{tidytext-package}
\title{tidytext: Text Mining using 'dplyr', 'ggplot2', and Other Tidy Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Using tidy data principles can make many text mining tasks easier, more effective, and consistent with tools already in wide use. Much of the infrastructure needed for text mining with tidy data frames already exists in packages like 'dplyr', 'broom', 'tidyr', and 'ggplot2'. In this package, we provide functions and supporting data sets to allow conversion of text to and from tidy formats, and to switch seamlessly between tidy tools and existing text mining packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://juliasilge.github.io/tidytext/}
  \item \url{https://github.com/juliasilge/tidytext}
  \item Report bugs at \url{https://github.com/juliasilge/tidytext/issues}
}

}
\author{
\strong{Maintainer}: Julia Silge \email{julia.silge@gmail.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID})

Authors:
\itemize{
  \item David Robinson \email{admiral.david@gmail.com}
}

Other contributors:
\itemize{
  \item Gabriela De Queiroz \email{gabidequeiroz@gmail.com} [contributor]
  \item Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID}) [contributor]
  \item Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} [contributor]
  \item Os Keyes \email{ironholds@gmail.com} (\href{https://orcid.org/0000-0001-5196-609X}{ORCID}) [contributor]
  \item Kanishka Misra \email{kmisra@purdue.edu} [contributor]
  \item Tim Mastny \email{tim.mastny@gmail.com} [contributor]
  \item Jeff Erickson \email{jeff@erick.so} [contributor]
}

}
\keyword{internal}
