% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytext.R
\docType{package}
\name{tidytext}
\alias{tidytext}
\title{tidytext: Text Mining using 'dplyr', 'ggplot2', and Other Tidy Tools}
\description{
This package implements tidy data principles to make many text mining tasks easier,
more effective, and consistent with tools already in wide use.
}
\details{
Much of the infrastructure needed for text mining with tidy data frames
already exists in packages like dplyr, broom, tidyr and ggplot2.

In this package, we provide functions and supporting data sets to allow conversion
of text to and from tidy formats, and to switch seamlessly between tidy tools
and existing text mining packages.

To learn more about tidytext, start with the vignettes:
\code{browseVignettes(package = "tidytext")}
}
