% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_ngrams.R
\name{unnest_ngrams}
\alias{unnest_ngrams}
\title{Wrapper around unnest_tokens & ngram}
\usage{
unnest_ngrams(tbl, output, input, n = 3L, n_min = n,
  stopwords = character(), ngram_delim = " ", simplify = FALSE,
  format = c("text", "man", "latex", "html", "xml"), to_lower = TRUE,
  drop = TRUE, collapse = NULL)
}
\arguments{
\item{tbl}{A data frame}

\item{output}{Output column to be created as string or symbol.}

\item{input}{Input column that gets split as string or symbol.

The output/input arguments are passed by expression and support
\link[rlang]{quasiquotation}; you can unquote strings and symbols.}

\item{n}{The number of words in the n-gram. This must be an integer greater
than or equal to 1.}

\item{n_min}{This must be an integer greater than or equal to 1, and less
than or equal to \code{n}.}

\item{stopwords}{A character vector of stop words to be excluded from the
n-grams.}

\item{ngram_delim}{The separator between words in an n-gram.}

\item{simplify}{\code{FALSE} by default so that a consistent value is
returned regardless of length of input. If \code{TRUE}, then an input with
a single element will return a character vector of tokens instead of a
list.}

\item{format}{Either "text", "man", "latex", "html", or "xml". If not text,
this uses the hunspell tokenizer, and can tokenize only by "word"}

\item{to_lower}{Whether to convert tokens to lowercase. If tokens include
URLS (such as with \code{token = "tweets"}), such converted URLs may no
longer be correct.}

\item{drop}{Whether original input column should get dropped. Ignored
if the original input and new output column have the same name.}

\item{collapse}{Whether to combine text with newlines first in case tokens
(such as sentences or paragraphs) span multiple lines. If NULL, collapses
when token method is "ngrams", "skip_ngrams", "sentences", "lines",
"paragraphs", or "regex".}
}
\description{
This function is a wrapper around \code{unnest_tokens( token = "ngrams" )}.
}
\examples{
library(dplyr)
library(janeaustenr)

d <- tibble(txt = prideprejudice)
d
d \%>\%
  unnest_ngrams(word, txt)

}
\seealso{
\itemize{
\item \code{\link[=unnest_tokens]{unnest_tokens()}}
}
}
