% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tokens.R
\name{unnest_tokens}
\alias{unnest_tokens}
\alias{unnest_tokens_}
\title{Split a column into tokens using either the tokenizer package or str_split}
\usage{
unnest_tokens_(tbl, output_col, input_col, token = "words", to_lower = TRUE,
  drop = TRUE, ...)

unnest_tokens(tbl, output, input, token = "words", to_lower = TRUE,
  drop = TRUE, ...)
}
\arguments{
\item{tbl}{Data frame}

\item{output_col}{Output column to be created}

\item{input_col}{Input column that gets split}

\item{token}{Unit for tokenizing. If tokenizer package
is installed, options are "characters", "words", "sentences", "lines",
"paragraphs", and "regex". Otherwise, using str_split the option is "words".
Default is "words".}

\item{to_lower}{Whether to turn column lowercase}

\item{drop}{Whether original input column should get dropped. Ignored
if the original input and new output column have the same name.}

\item{...}{Extra arguments passed on to the tokenizer}

\item{output}{Output column to be created as bare name}

\item{input}{Input column that gets split as bare name}
}
\description{
Split a column into tokens using either the tokenizer package or str_split
}
\details{
If the unit for tokenizing is sentences, lines, paragraphs, or
regex, the entire input will be collapsed together before tokenizing.
}
\examples{

library(dplyr)
library(janeaustenr)

d <- data_frame(txt = prideprejudice)
d

d \%>\%
  unnest_tokens(word, txt)

d \%>\%
  unnest_tokens(word, txt, token = "sentences")

}

