% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tidytext.R
\name{bind_tf_idf_}
\alias{bind_tf_idf_}
\alias{cast_sparse_}
\alias{cast_tdm_}
\alias{cast_dtm_}
\alias{cast_dfm_}
\alias{unnest_tokens_}
\alias{deprecated-se}
\title{Deprecated SE version of functions}
\usage{
bind_tf_idf_(...)

cast_sparse_(...)

cast_tdm_(...)

cast_dtm_(...)

cast_dfm_(...)

unnest_tokens_(...)
}
\arguments{
\item{...}{All arguments for non-deprecated functions}
}
\description{
tidytext used to offer twin versions of each verb suffixed with an
underscore, like dplyr and the main tidyverse packages. These
versions had standard evaluation (SE) semantics; rather than taking
arguments by code, like NSE verbs, they took arguments by value.
Their purpose was to make it possible to program with tidytext.
However, tidytext now uses tidy evaluation semantics. NSE verbs
still capture their arguments, but you can now unquote parts of
these arguments. This offers full programmability with NSE verbs.
Thus, the underscored versions are now superfluous.
}
\details{
Unquoting triggers immediate evaluation of its operand and inlines
the result within the captured expression. This result can be a
value or an expression to be evaluated later with the rest of the
argument.
}
\keyword{internal}
