% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_tidiers.R
\name{dictionary_tidiers}
\alias{dictionary_tidiers}
\alias{tidy.dictionary2}
\title{Tidy dictionary objects from the quanteda package}
\usage{
\method{tidy}{dictionary2}(x, regex = FALSE, ...)
}
\arguments{
\item{x}{A dictionary object}

\item{regex}{Whether to turn dictionary items from a glob to a regex}

\item{...}{Extra arguments, not used}
}
\value{
A data frame with two columns: category and word.
}
\description{
Tidy dictionary objects from the quanteda package
}
