% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_tidiers.R, R/lda_tidiers.R,
%   R/sparse_tidiers.R, R/stm_tidiers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{glance}
\alias{augment}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}}

