% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiments.R
\docType{data}
\name{sentiments}
\alias{sentiments}
\title{Sentiment lexicon from Bing Liu and collaborators}
\format{
A data frame with 6,786 rows and 2 variables:
\describe{
\item{word}{An English word}
\item{sentiment}{A sentiment for that word, either positive or negative.}
}
}
\source{
\url{https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\usage{
sentiments
}
\description{
Lexicon for opinion and sentiment analysis in a tidy data frame. This
dataset is included in this package with permission of the creators, and
may be used in research, commercial, etc. contexts with attribution, using
either the paper or URL below.
}
\details{
This lexicon was first published in:

Minqing Hu and Bing Liu, ``Mining and summarizing customer reviews.'',
Proceedings of the ACM SIGKDD International Conference on Knowledge Discovery
& Data Mining (KDD-2004), Seattle, Washington, USA, Aug 22-25, 2004.

Words with non-ASCII characters were removed.
}
\keyword{datasets}
