% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Various lexicons for English stop words}
\format{A data frame with 1149 rows and 2 variables:
\describe{
 \item{word}{An English word}
 \item{lexicon}{The source of the stop word. Either "onix", "SMART", or "snowball"}
 }}
\source{
\url{http://www.lextek.com/manuals/onix/stopwords1.html}
\url{http://jmlr.csail.mit.edu/papers/volume5/lewis04a/a11-smart-stop-list/english.stop}
\url{http://snowball.tartarus.org/algorithms/english/stop.txt}
}
\usage{
stopwords
}
\description{
English stop words from three lexicons, as a data frame.
The onix and SMART sets are pulled from the tm package. Note
that words with non-ASCII have been removed. The raw data is in
\code{inst/extdata}.
}
\seealso{
\code{\link{setup_data}}
}
\keyword{datasets}

