% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tokens.R
\name{unnest_tokens}
\alias{unnest_tokens}
\alias{unnest_tokens_}
\title{Split a column into tokens}
\usage{
unnest_tokens_(tbl, token_col, text_col, method = NULL, to_lower = TRUE,
  drop = TRUE, ...)
}
\arguments{
\item{tbl}{Data frame}

\item{token_col}{Token column to be created}

\item{text_col}{Text column that gets split}

\item{method}{Method for tokenizing. If tokenizer package
is installed, use tokenize. Otherwise, use str_split.}

\item{to_lower}{Whether to turn column lowercase}

\item{drop}{Whether original text column should get dropped}

\item{...}{Extra arguments passed on to the tokenizer}
}
\description{
Split a column into tokens
}

