% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tokens.R
\name{unnest_tokens}
\alias{unnest_tokens}
\alias{unnest_tokens_}
\title{Split a column into tokens using either the tokenizer package or str_split}
\usage{
unnest_tokens_(tbl, token_col, text_col, method = NULL, to_lower = TRUE,
  drop = TRUE, ...)

unnest_tokens(tbl, token, text, method = NULL, to_lower = TRUE,
  drop = TRUE, ...)
}
\arguments{
\item{tbl}{Data frame}

\item{token_col}{Token column to be created}

\item{text_col}{Text column that gets split}

\item{method}{Method for tokenizing. If tokenizer package
is installed, options are "characters", "words", "sentences", "lines",
"paragraphs", and "regex". Otherwise, using str_split the option is "words".
Default is "words".}

\item{to_lower}{Whether to turn column lowercase}

\item{drop}{Whether original text column should get dropped}

\item{...}{Extra arguments passed on to the tokenizer}
}
\description{
Split a column into tokens using either the tokenizer package or str_split
}

