% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_tweets.R
\name{unnest_tweets}
\alias{unnest_tweets}
\title{Wrapper around unnest_tokens for tweets}
\usage{
unnest_tweets(tbl, output, input, ...)
}
\arguments{
\item{tbl}{A data frame}

\item{output}{Output column to be created as string or symbol.}

\item{input}{Input column that gets split as string or symbol.

The output/input arguments are passed by expression and support
\link[rlang:topic-inject]{quasiquotation}; you can unquote strings and symbols.}

\item{...}{Extra arguments passed on to \link[tokenizers:tokenizers]{tokenizers}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\seealso{
\itemize{
\item \code{\link[=unnest_tokens]{unnest_tokens()}}
}
}
\keyword{internal}
